Factor out HttpClient inner interfaces
