+++ azure-client-runtime/src/test/java/com/microsoft/azure/v2/AzureProxyToRestProxyWithNettyTests.java
@@ -1,17 +1,14 @@
 package com.microsoft.azure.v2;
 
 import com.microsoft.rest.v2.http.HttpClient;
-import com.microsoft.rest.v2.http.HttpClient.Configuration;
+import com.microsoft.rest.v2.http.HttpClientConfiguration;
 import com.microsoft.rest.v2.http.NettyClient;
-import com.microsoft.rest.v2.policy.RequestPolicy;
-
-import java.util.Collections;
 
 public class AzureProxyToRestProxyWithNettyTests extends AzureProxyToRestProxyTests {
     private final NettyClient.Factory nettyClientFactory = new NettyClient.Factory();
 
     @Override
     protected HttpClient createHttpClient() {
-        return nettyClientFactory.create(new Configuration(null));
+        return nettyClientFactory.create(new HttpClientConfiguration(null));
     }
 }
