+++ client-runtime/src/main/java/com/microsoft/rest/v2/http/HttpClient.java
@@ -6,10 +6,7 @@
 
 package com.microsoft.rest.v2.http;
 
-import io.reactivex.Completable;
 import io.reactivex.Single;
-
-import java.net.Proxy;
 
 /**
  * A generic interface for sending HTTP requests and getting responses.
@@ -25,7 +22,7 @@ public abstract class HttpClient {
     private static final class DefaultHttpClientHolder {
         // Putting this field in an inner class makes it so it is only instantiated when
         // one of the createDefault() methods instead of instantiating when any members are accessed.
-        private static HttpClient.Factory defaultHttpClientFactory = new NettyClient.Factory();
+        private static HttpClientFactory defaultHttpClientFactory = new NettyClient.Factory();
     }
 
     /**
@@ -41,50 +38,7 @@ public abstract class HttpClient {
      * @param configuration The configuration to apply to the HttpClient.
      * @return an instance of the default HttpClient type.
      */
-    public static HttpClient createDefault(HttpClient.Configuration configuration) {
+    public static HttpClient createDefault(HttpClientConfiguration configuration) {
         return DefaultHttpClientHolder.defaultHttpClientFactory.create(configuration);
-    }
-
-    /**
-     * The set of parameters used to create an HTTP client.
-     */
-    public static final class Configuration {
-        private final Proxy proxy;
-
-        /**
-         * @return The optional proxy to use.
-         */
-        public Proxy proxy() {
-            return proxy;
-        }
-
-        /**
-         * Creates a Configuration.
-         * @param proxy The optional proxy to use.
-         */
-        public Configuration(Proxy proxy) {
-            this.proxy = proxy;
-        }
-    }
-
-    /**
-     * Creates an HttpClient from a Configuration.
-     */
-    public interface Factory {
-        /**
-         * Creates an HttpClient with the given Configuration.
-         * @param configuration the configuration.
-         * @return the HttpClient.
-         */
-        HttpClient create(Configuration configuration);
-
-        /**
-         * Asynchronously awaits completion of in-flight tasks,
-         * then closes shared resources associated with this HttpClient.Factory.
-         * After this Completable completes, HttpClients created from this Factory can no longer be used.
-         *
-         * @return a Completable which shuts down the factory when subscribed to.
-         */
-        Completable shutdown();
     }
 }
