+++ client-runtime/src/test/java/com/microsoft/rest/v2/RestProxyWithHttpProxyNettyTests.java
@@ -1,13 +1,12 @@
 package com.microsoft.rest.v2;
 
 import com.microsoft.rest.v2.http.HttpClient;
+import com.microsoft.rest.v2.http.HttpClientConfiguration;
 import com.microsoft.rest.v2.http.NettyClient;
-import com.microsoft.rest.v2.policy.RequestPolicy;
 import org.junit.Ignore;
 
 import java.net.InetSocketAddress;
 import java.net.Proxy;
-import java.util.Collections;
 
 @Ignore("Should only be run manually when a local proxy server (e.g. Fiddler) is running")
 public class RestProxyWithHttpProxyNettyTests extends RestProxyTests {
@@ -17,6 +16,6 @@ public class RestProxyWithHttpProxyNettyTests extends RestProxyTests {
     protected HttpClient createHttpClient() {
         InetSocketAddress address = new InetSocketAddress("127.0.0.1", 8888);
         Proxy proxy = new Proxy(Proxy.Type.HTTP, address);
-        return nettyClientFactory.create(new HttpClient.Configuration(proxy));
+        return nettyClientFactory.create(new HttpClientConfiguration(proxy));
     }
 }
